+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v1.0.1
+- Add conditionally formatting macros: `formatmc`, `printmc`, `printmcln`,
+  `eprintmc`, `eprintmcln`
+
 ## v1.0.0
 - Fix inline variable names in the `*c` macros
 - New formatting macros: `eprintc`, `eprintcln`